refactor: update import paths for confirm update services and entities
